Fix bug: Some evaluators assign bg probability scores
QL Evaluators like Jelenik-Mercer or Dirichlet score even
those documents that don't contain a query-term, by
assigning them a background probability for the query-term.
So, just checking if an inverted-list has a posting for a
doc-id is not enough to decide if the doc can be scored. It
could be scored w.r.t a query term even if it doesn't contain
it, provided the evaluator accepts such a document.

To this end, add an Evaluator.assignsBackgroundProbability()
API which tells the retriver whether to send a doc to the
evaluator when a query-term doesn't appear in it.